Added static Utillities class
